requirements.txt
opencv-python
numpy
packaging
setuptools
pillow
lightning
scikit-learn
timm
wandb
matplotlib
einops
albumentations
PyYAML
seaborn
adabelief-pytorch
torchmetrics